Fix bugs in ROOT conversion script
There were two bugs
    - The first event was always empty
    - The latter half of channels always had no waveform data (all
      zeros)

This commit fixes both of those bugs. The first was caused by the loop
count variable being incremented at the start of the loop instead of the
end. So the first loop had a value of 1 not 0 in that variable.

The second was caused by using the "nsample" variable from CERES header,
which is actually half the number of samples, so needs to be doubled
before being used.